+++ app/dashboard/(overview)/loading.tsx
@@ -1,0 +1,7 @@
+import DashboardSkeleton from '@/app/ui/skeletons';
+
+export default function Loading() {
+    return <DashboardSkeleton />
+}
+
+//* loading.tsx file will only apply to dashboard overview page.
